chore(#541): moved mnist-converter
The function to convert MNIST-files into datasets
was moved for better code-separation.
